pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=6.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "i10-bluesky"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
description = ""
dependencies = [
    "typing-extensions;python_version<'3.8'",

    "pyqt5",
    "setuptools",
    "networkx>=2.0",
    "numpy",
    "packaging",
    "pint",
    "event-model",
    "p4p",
    "pyyaml",
] # Add project dependencies here, e.g. ["click", "numpy"]
dynamic = ["version"]
#version = "0.0.1"         #need to make this non manual but dynamic version currently include local version seed and it not the statand for release. 
license.file = "LICENSE"
readme = "README.rst"
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "bluesky",
    "ophyd-async",
    "black",
    "mypy",
    "pipdeptree",
    "pre-commit",
    "pydata-sphinx-theme>=0.12",
    "pytest",
    "pytest-cov",
    "ruff",
    "sphinx-autobuild",
    "sphinx-copybutton",
    "sphinx-design",
    "tox-direct",
    "types-mock",
    "copier",
    "ophyd_async[pva]",
    "ophyd_async[ca]",
    "flake8",
    "flake8-isort",
    "Flake8-pyproject",
    "h5py",
    "inflection",
    "ipython",
    "ipywidgets",
    "matplotlib",
    "myst-parser",
    "numpydoc",
    "ophyd",
    "pickleshare",
    "pyepics>=3.4.2",
    "pyside6",
    "pytest-asyncio",
    "pytest-faulthandler",
    "pytest-rerunfailures",
    "pytest-timeout",
    "types-pyyaml",
    "jupyter",
    "softioc",
    "cothread",
    "dls-dodal",

]

[project.scripts]
i10-bluesky = "i10_bluesky.__main__:main"

[project.urls]
GitHub = "https://github.com/Relm-Arrowny/i10-bluesky"

[[project.authors]] # Further authors may be added by duplicating this section
email = "raymond.fan@diamond.ac.uk"
name = "Raymond Fan"


[tool.setuptools_scm]
write_to = "src/i10_bluesky/_version.py"

[tool.mypy]
ignore_missing_imports = true # Ignore missing stubs in imported modules

[tool.pytest.ini_options]
# Run pytest with all our checkers, and don't spam us with massive tracebacks on error
addopts = """
    --tb=native -vv --doctest-modules --doctest-glob="*.rst"
    """
# https://iscinumpy.gitlab.io/post/bound-version-constraints/#watch-for-warnings
filterwarnings = "error"
# Doctest python code in docs, python code in src docstrings, test functions in tests
testpaths = "docs src tests"

asyncio_mode = "auto"

[tool.coverage.run]
data_file = "/tmp/i10_bluesky.coverage"

[tool.coverage.paths]
# Tests are run from installed location, map back to the src directory
source = ["src", "**/site-packages/"]

# tox must currently be configured via an embedded ini string
# See: https://github.com/tox-dev/tox/issues/999
[tool.tox]
legacy_tox_ini = """
[tox]
skipsdist=True

[testenv:{pre-commit,mypy,pytest,docs}]
# Don't create a virtualenv for the command, requires tox-direct plugin
direct = True
passenv = *
allowlist_externals =
    pytest
    pre-commit
    mypy
    sphinx-build
    sphinx-autobuild
commands =
    pytest: pytest --cov=i10_bluesky --cov-report term --cov-report xml:cov.xml {posargs}
    mypy: mypy src tests {posargs}
    pre-commit: pre-commit run --all-files {posargs}
    docs: sphinx-{posargs:build -EW --keep-going} -T docs build/html
"""


[tool.ruff]
src = ["src", "tests"]
line-length = 88
select = [
    "C4",   # flake8-comprehensions - https://beta.ruff.rs/docs/rules/#flake8-comprehensions-c4
    "E",    # pycodestyle errors - https://beta.ruff.rs/docs/rules/#error-e
    "F",    # pyflakes rules - https://beta.ruff.rs/docs/rules/#pyflakes-f
    "W",    # pycodestyle warnings - https://beta.ruff.rs/docs/rules/#warning-w
    "I001", # isort
]
